reorganized into a Python package
