+++ main.py
@@ -1,10 +1,11 @@
-import config_parser
 import numpy as np
 import os
 import time
 import torch
 import tqdm
-import utils
+
+from nerf_shared import config_parser
+from nerf_shared import utils
 from torch.utils.tensorboard.writer import SummaryWriter
 
 device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
@@ -121,20 +122,20 @@ def run():
             if i%args.i_video==0 and i > 0:
                 utils.render_training_video(args, render_poses, hwf, K, render_kwargs_test, i)
             '''
-            # Renders out the test poses (i.e. poses[i_test]) to visually evaluate NeRF quality 
+            # Renders out the test poses (i.e. poses[i_test]) to visually evaluate NeRF quality
             if i%args.i_testset==0 and i > 0:
-                renderer.render_from_batch_poses(H, 
-                                                 W, 
-                                                 K, 
-                                                 chunk=args.chunk, 
-                                                 batch_c2w=poses[i_test], 
-                                                 coarse_model=coarse_model, 
-                                                 fine_model=fine_model, 
-                                                 retraw=True, 
+                renderer.render_from_batch_poses(H,
+                                                 W,
+                                                 K,
+                                                 chunk=args.chunk,
+                                                 batch_c2w=poses[i_test],
+                                                 coarse_model=coarse_model,
+                                                 fine_model=fine_model,
+                                                 retraw=True,
                                                  save_directory=os.path.join(args.basedir, args.expname, 'testset_{:06d}'.format(i)),
                                                  b_combine_as_video=False,
                                                  tb_writer=tb_writer)
-            
+
             #Displays loss and PSNR (Peak signal to noise ratio) of the fine reconstruction loss
             if i%args.i_print==0:
                 utils.print_statistics(args, loss, psnr, i, tb_writer=tb_writer)
@@ -146,7 +147,7 @@ def run():
         pass
 
 if __name__=='__main__':
-    if device.type != 'cpu': 
+    if device.type != 'cpu':
         torch.set_default_tensor_type('torch.cuda.FloatTensor')
         torch.cuda.empty_cache()
