+++ nerf_shared/load_LINEMOD.py
