+++ nerf_shared/load_blender.py
@@ -1,7 +1,7 @@
 import os
 import torch
 import numpy as np
-import imageio 
+import imageio
 import json
 import torch.nn.functional as F
 import cv2
@@ -44,6 +44,7 @@ def pose_spherical(theta, phi, radius):
 def load_blender_data(basedir, half_res=False, testskip=1):
     splits = ['train', 'val', 'test']
     metas = {}
+    print(os.getcwd())
     for s in splits:
         with open(os.path.join(basedir, 'transforms_{}.json'.format(s)), 'r') as fp:
             metas[s] = json.load(fp)
@@ -59,7 +60,7 @@ def load_blender_data(basedir, half_res=False, testskip=1):
             skip = 1
         else:
             skip = testskip
-            
+
         for frame in meta['frames'][::skip]:
             fname = os.path.join(basedir, frame['file_path'] + '.png')
             imgs.append(imageio.imread(fname))
@@ -69,18 +70,18 @@ def load_blender_data(basedir, half_res=False, testskip=1):
         counts.append(counts[-1] + imgs.shape[0])
         all_imgs.append(imgs)
         all_poses.append(poses)
-    
+
     i_split = [np.arange(counts[i], counts[i+1]) for i in range(3)]
-    
+
     imgs = np.concatenate(all_imgs, 0)
     poses = np.concatenate(all_poses, 0)
-    
+
     H, W = imgs[0].shape[:2]
     camera_angle_x = float(meta['camera_angle_x'])
     focal = .5 * W / np.tan(.5 * camera_angle_x)
-    
+
     render_poses = torch.stack([pose_spherical(angle, 0, 4.0) for angle in np.linspace(-180,180,40+1)[:-1]], 0)
-    
+
     if half_res:
         H = H//2
         W = W//2
@@ -92,7 +93,7 @@ def load_blender_data(basedir, half_res=False, testskip=1):
         imgs = imgs_half_res
         # imgs = tf.image.resize_area(imgs, [400, 400]).numpy()
 
-        
+
     return imgs, poses, render_poses, [H, W, focal], i_split
