+++ nerf_shared/nerf.py
