+++ setup.py
@@ -1,0 +1,3 @@
+from setuptools import setup
+
+setup(name='nerf_shared', version='1.0', packages=['nerf_shared'])
